Move Platform enum naming into Core/Serialization
The names we give PlatformKey and PlatformWindowMode values are ours, not the
platform layer's. PlatformKey lived in Io next to the InputAction code under a
TODO asking for a better place, and PlatformWindowMode sat in GameConfig.

The PlatformKey one was the worse of the two - it made Core/Trace depend on Io,
which is backwards, since Core is only allowed to reach into Utility and
Library. The dependency was never declared either, so core_trace compiled
against an Io header and picked up the symbols from whatever else the final
binary happened to link.

Core/Serialization is where engine-specific serialization code goes, the
counterpart to Library/Serialization. Io keeps everything that is really about
InputAction.

This takes Core/Trace out of the big dependency cycle, which is down to 26
packages from 27.
