+++ src/Core/CMakeLists.txt
@@ -1,3 +1,4 @@
 cmake_minimum_required(VERSION 3.27 FATAL_ERROR)
 
+add_subdirectory(Serialization)
 add_subdirectory(Trace)
