+++ src/Core/Serialization/PlatformKeySerialization.h
@@ -1,0 +1,12 @@
+#pragma once
+
+#include <string>
+#include <string_view>
+
+#include "Library/Platform/Interface/PlatformEnums.h"
+#include "Library/Serialization/SerializationFwd.h"
+
+MM_DECLARE_SERIALIZATION_FUNCTIONS(PlatformKey)
+
+std::string GetDisplayName(PlatformKey key);
+bool TryParseDisplayName(std::string_view displayName, PlatformKey *outKey);
