+++ src/Core/Serialization/PlatformWindowModeSerialization.cpp
@@ -1,0 +1,10 @@
+#include "PlatformWindowModeSerialization.h"
+
+#include "Library/Serialization/EnumSerialization.h"
+
+MM_DEFINE_ENUM_SERIALIZATION_FUNCTIONS(PlatformWindowMode, CASE_INSENSITIVE, {
+    {WINDOW_MODE_WINDOWED, "windowed"},
+    {WINDOW_MODE_BORDERLESS, "borderless"},
+    {WINDOW_MODE_FULLSCREEN_BORDERLESS, "fullscreen_borderless"},
+    {WINDOW_MODE_FULLSCREEN_BORDERLESS, "borderless_fullscreen"}, // Alias for convenience
+});
