+++ src/Core/Serialization/PlatformWindowModeSerialization.h
@@ -1,0 +1,6 @@
+#pragma once
+
+#include "Library/Platform/Interface/PlatformEnums.h"
+#include "Library/Serialization/SerializationFwd.h"
+
+MM_DECLARE_SERIALIZATION_FUNCTIONS(PlatformWindowMode)
