+++ src/Core/Trace/CMakeLists.txt
@@ -12,6 +12,7 @@ set(CORE_TRACE_HEADERS
 add_library(core_trace STATIC ${CORE_TRACE_SOURCES} ${CORE_TRACE_HEADERS})
 target_check_style(core_trace)
 target_link_libraries(core_trace PUBLIC
+        core_serialization
         library_serialization
         library_json
         library_platform_interface
