+++ src/Io/InputEnumFunctions.h
@@ -1,10 +1,6 @@
 #pragma once
 
 #include <string>
-#include <string_view>
-
-#include "Library/Platform/Interface/PlatformEnums.h"
-#include "Library/Serialization/SerializationFwd.h"
 
 #include "Utility/Segment.h"
 
@@ -21,9 +17,3 @@ inline Segment<InputAction> allConfigurableInputActions() {
 std::string GetDisplayName(InputAction action);
 
 InputActionTriggerMode triggerModeForInputAction(InputAction action);
-
-// TODO(captainurist): find a better place for this code
-MM_DECLARE_SERIALIZATION_FUNCTIONS(PlatformKey)
-
-std::string GetDisplayName(PlatformKey key);
-bool TryParseDisplayName(std::string_view displayName, PlatformKey *outKey);
